[SEARCH-FRIENDS] Split up friends logic in list and search
